+++ examples/data-sources/meraki_appliance_site_to_site_vpn/data-source.tf
@@ -1,0 +1,3 @@
+data "meraki_appliance_site_to_site_vpn" "example" {
+  network_id = "L_123456"
+}