+++ examples/resources/meraki_appliance_site_to_site_vpn/import.sh
@@ -1,0 +1,1 @@
+terraform import meraki_appliance_site_to_site_vpn.example "<network_id>"