+++ CHANGELOG.md
@@ -36,6 +36,7 @@
 - Add `meraki_appliance_traffic_shaping_uplink_selection` resource and data source
 - Add `meraki_appliance_traffic_shaping_vpn_exclusions` resource and data source
 - Add `meraki_appliance_uplink_settings` resource and data source
+- Add `meraki_appliance_site_to_site_vpn` resource and data source
 
 ## 0.1.1